refactor: types/validation schema in separate files + change base url
