+++ React Project/src/configs/urls.ts
@@ -1,2 +1,1 @@
-export const BASE_URL: string = "https://sheetlabs.com/CRES/project_list";
-export const SERVER_URL: string = "http://localhost:5000";
+export const BASE_URL: string = "http://localhost:5000";
