+++ React Project/src/lib/axios.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
-import { SERVER_URL } from "../configs/urls";
+import { BASE_URL } from "../configs/urls";
 
 const axiosInstance = axios.create({
-    baseURL: SERVER_URL,
+    baseURL: BASE_URL,
 });
 
 export default axiosInstance;
